adds access modifiers and sharpImageProcessor refactor
