+++ src/colors/imageProcessing/index.ts
@@ -1,0 +1,3 @@
+import SharpImageProcessor  from './sharpImageProcessor';
+
+export default SharpImageProcessor;
